Diagram consistency
The diagrams were inconsistent, so I went through and lined everything
up.  Some of the diagrams were generating warnings about being too wide,
which is what triggered this.

All the client messages start on column 0 now.  All the server messages
end on column 55 (this is based on getting wider images with the flight
labels within 70 columns).  All the arrows now go from column 23 to 32,
based on the location in the above.

All figures have consistent headers for client and server.